requirements.txt
fastapi==0.112.2
httpx==0.27.2
motor==3.5.1
numpy==2.1.0
pandas==2.2.2
pydantic==2.8.2
PyJWT==2.9.0
pymongo==4.8.0
Requests==2.32.3
starlette==0.38.2
uvicorn==0.30.6
